make Ising critical temperature a `BigFloat`

+++ src/models/ising.jl
@@ -1,4 +1,4 @@
-const Ising_βc = log(1.0 + sqrt(2)) / 2.0
+const Ising_βc = BigFloat(log(1.0 + sqrt(2)) / 2.0)
 Potts_βc(q) = log(1.0 + sqrt(q))
 
 const f_onsager::BigFloat = -2.10965114460820745966777928351108478082549327543540531781696107967700291143188081390114126499095041781
